backend/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.0.0
livekit-agents~=1.4
livekit-plugins-openai~=1.4
livekit-plugins-silero~=1.4
livekit-api>=1.0.0
